[CHG][NT] 修复使用 16 bit 数据存储 AD 转换的结果

+++ bsp/gd32/gd32205/drivers/drv_tcd1209.c
@@ -18,7 +18,7 @@
 #define DBG_TAG    "tcd1209"
 #include <rtdbg.h>
 
-static uint8_t gs_ad9945_data[AD9945_DATA_COUNTS];
+static uint16_t gs_ad9945_data[AD9945_DATA_COUNTS];
 /*
  * AHB = 120M
  * APB2 = 120M
